ocr_engine/requirements.txt
# Core ML / OCR
paddleocr>=2.7.0
paddlepaddle>=2.6.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
datasets>=2.16.0
accelerate>=0.27.0
seqeval>=1.2.2

# Image / PDF handling
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pdf2image>=1.17.0
scikit-image>=0.22.0
numpy>=1.26.0

# Synthetic data generation
Faker>=24.0.0
reportlab>=4.0.0
jinja2>=3.1.0

# Utilities
pydantic>=2.6.0
tqdm>=4.66.0
PyYAML>=6.0.1
requests>=2.31.0
click>=8.1.0

# Experiment tracking (optional)
mlflow>=2.10.0

# Dev / test
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.3.0
